catalog_pipeline/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
faiss-cpu==1.8.0
numpy==1.26.4
pymupdf==1.24.9
pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
packaging==24.2
